# Add TORCH_CHECK_INDEX in convert_indices_from_coo_to_csr_cpu