@cpuhrsch 2024-10-16
@Kh4L - have you tried using the `check_invariants` flag of [torch.sparse_coo_tensor](https://pytorch.org/docs/main/generated/torch.sparse_coo_tensor.html#torch.sparse_coo_tensor)?