@Kh4L 2024-10-17
> @Kh4L - have you tried using the `check_invariants` flag of [torch.sparse_coo_tensor](https://pytorch.org/docs/main/generated/torch.sparse_coo_tensor.html#torch.sparse_coo_tensor)?

Yes, I gave it a try
However, I still believe it’s worth checking the index here, as we were encountering various types of memory corruption issues that were causing Python to crash, making it difficult to debug

I think the cost of adding this index check is relatively small compared to the time it could save developers in the long run