@cpuhrsch 2024-10-17
@Kh4L - I agree, I'd have hoped that `check_invariants` and the [`torch.sparse.check_sparse_tensor_invariants`](https://pytorch.org/docs/main/generated/torch.sparse.check_sparse_tensor_invariants.html#check-sparse-tensor-invariants) context manager already did all the relevant checks needed here. Did you try running under `with torch.sparse.check_sparse_tensor_invariants():` as in
```
with torch.sparse.check_sparse_tensor_invariants():
    run_my_model()
```