@cpuhrsch 2024-10-17
If this doesn't work, I suggest we modify this PR to use https://pytorch.org/docs/main/generated/torch.sparse.check_sparse_tensor_invariants.html#torch.sparse.check_sparse_tensor_invariants.is_enabled as a guard. That way this can be turned on for debugging, but doesn't affect performance by default.