@Kh4L 2024-10-17
@cpuhrsch I understand but is it really ok to have python crashing with errors such as `malloc(): invalid next size (unsorted)` even without `check_sparse_tensor_invariants` ? `to_sparse_csr` does not seem to be a typical bottleneck