@Kh4L 2024-10-17
@cpuhrsch 
the index check if optimized and the cost is negligible, see it for a relatively big sptensor:

```python
import torch
import time
import numpy as np

num_nonzeros = 2 ** 14
dense_size = (2**16, 2**14)

row_indices = torch.randint(0, dense_size[0], (num_nonzeros,))
col_indices = torch.randint(0, dense_size[1], (num_nonzeros,))
coo_indices = torch.stack((row_indices, col_indices))
sparse_coo = torch.sparse_coo_tensor(
    coo_indices,
    torch.ones(coo_indices.size(1), dtype=bool),
    dense_size,
)

def benchmark_to_sparse_csr():
    return sparse_coo.to_sparse_csr()

num_runs = 100
times = []

for _ in range(num_runs):
    start_time = time.time()
    benchmark_to_sparse_csr()
    end_time = time.time()
    times.append(end_time - start_time)

print(f"Mean execution time: {np.mean(times):.6f} seconds")
print(f"Standard deviation: {np.std(times):.6f} seconds")
```
before the change (without index check):
```
Mean execution time: 0.178715 seconds
Standard deviation: 0.004270 seconds
```
after the change (with index check):
```
Mean execution time: 0.178128 seconds
Standard deviation: 0.003673 seconds
```