@cpuhrsch 2024-10-17
@Kh4L - when the Tensor is big it's fine, but maybe when it's small it's not fine anymore for some applications. By supporting the guard the user can choose what works best for them.